# Optimize UDF with parallel execution