## Highlights

In this PR I am:

- Passing IDs only instead of whole rows into parallel processes UDFs. This will prevent multiple data conversion (pickling) along the way
- Write to DB right from parallel process instead of passing results pack into main process

This will prevent multiple types conversion.

Before it was:

1. Read rows from DB -> convert them from raw DB types into Python types (whole raw)
2. Pass rows into parallel processes via `multiprocess.Queue` -> convert rows from Python types with msgpack
3. Read rows in parallel process from `multiprocess.Queue` -> convert rows to Python types (maspack)
4. Process UDF
5. Send rows back to the main process via `multiprocess.Queue` -> convert from Python type (msgpack)
6. In main process read rows from `multiprocess.Queue` -> convert them back into Python type (msgpack)
7. Write rows into DB -> convert them from Python types into DB raw types

After:

1. Read rows IDs from DB -> convert DB `int`s into Python `int`s is quick, stable, predictable
2. Pass rows IDs to parallel processes in batches via `multiprocess.Queue` -> convert list of IDs with pickle (by default)
3. In parallel process for each batch of IDs read rows from DB -> convert them from raw DB types into Python types
4. Process UDF
5. Write result rows into DB right from the parallel process -> convert them from Python types into DB raw types

In the end:

- no multiple type conversion
- no more single point reading/writing from/to warehouse DB: read and write from parallel processes (worker machines in future)


## Test scenario

Simple script to check raw `parallel` setting only:

```python
import sys
from datachain import DataChain, File


def path_len(file: File) -> int:
    return len(file.path)


parallel = int(sys.argv[1]) if len(sys.argv) > 1 else None


DataChain \
    .from_dataset("50k-laion-files") \
    .settings(parallel=parallel, prefetch=0) \
    .map(path_len=path_len) \
    .save("50k-laion-files-len")
```

This is very simple and basic scenario, but it helps us to test `parallel` setting only, without any overheads.

**Note** `prefetch` is off in this case to measure `parallel` only


## Overview

On the chart below there are two series: before optimization (blue) and after (green). On the X axis is parallel processes count, on the Y axis is number of total rows processed by UDF in parallel. This is valid for SQLite warehouse on my local machine.

<img width="804" alt="image" src="https://github.com/user-attachments/assets/5e63efad-5fdd-46fe-96bb-988848e3bd2f" />

As we can see, "before" `parallel` option does not affect performance at all, there is a strict limit on performance and it does not depends on number of parallel processes.

The reason is because we pass rows into UDF for each parallel process via `multiprocess.Queue` and get results back the same way. `Queue` performance is very limited. I wrote [a simple script](https://gist.github.com/dreadatour/59508ce92010ebe6b4bc5ee97e430d21) to test `Queue` only and it is limited indeed. I have tried different ways for IPC (Pipes, ZeroMQ) and they all have this limit. This can be solved introducing external dependencies (Redis, RabbitMQ, etc), but it is not what we want for CLI tool.

"After" performs much better, `Queue` is used only to pass IDs in batches and is performant enough to show the performance boost depending on parallel processes count. It is not linear, because performance of SQLite warehouse is now the limit, but it is much better, stable and predictable.

Also note "1 parallel process" performance is ~2.15 times slower than clean "no parallel processes" and this is basically overhead for using parallel processes and queues to read and pass IDs. on 2-3 parallel processes performance is the same as on "no parallel" and it is increasing over parallel processes count increasing.

**Next** I am going to measure the same numbers on ClickHouse DB warehouse, I suppose it is going to be much better and linear.


## More measurements for those who love raw numbers

### Before

##### Not parallel (for reference)

```
$ time python path-len.py
Preparing: 129136 rows [00:01, 117496.00 rows/s]
Processed: 129136 rows [00:04, 30999.24 rows/s]
Cleanup: 2 tables [00:00, 297.27 tables/s]
python path-len.py  7.14s user 3.01s system 131% cpu 7.728 total
```

##### Parallel = 1 (edge case)

```
$ time python path-len.py 1
Preparing: 129136 rows [00:01, 114113.84 rows/s]
Processed: 129136 rows [00:11, 11689.08 rows/s]
Cleanup: 2 tables [00:00, 310.41 tables/s]
python path-len.py 1  20.59s user 11.62s system 202% cpu 15.872 total
```

##### Parallel = 8

```
$ time python path-len.py 8
Preparing: 129136 rows [00:01, 117677.17 rows/s]
Processed: 129136 rows [00:10, 12675.34 rows/s]
Cleanup: 2 tables [00:00, 331.29 tables/s]
python path-len.py 8  32.53s user 15.31s system 306% cpu 15.630 total
```

### After

##### Not parallel (for reference)

```
$ time python path-len.py
Preparing: 129136 rows [00:01, 113141.62 rows/s]
Processed: 129136 rows [00:04, 30850.59 rows/s]
Cleanup: 2 tables [00:00, 275.74 tables/s]
python path-len.py  7.06s user 3.38s system 135% cpu 7.701 total
```

##### Parallel = 1 (edge case)

```
$ time python path-len.py 1
Preparing: 129136 rows [00:01, 115298.96 rows/s]
Processed: 129137 rows [00:08, 15122.98 rows/s]
Cleanup: 2 tables [00:00, 317.81 tables/s]
python path-len.py 1  14.27s user 9.30s system 173% cpu 13.569 total
```

##### Parallel = 8

```
$ time python path-len.py 8
Preparing: 129136 rows [00:01, 115127.01 rows/s]
Processed: 129144 rows [00:03, 39761.77 rows/s]
Cleanup: 2 tables [00:00, 295.43 tables/s]
python path-len.py 8  27.65s user 13.99s system 452% cpu 9.206 total
```