@dreadatour 2024-12-14
> Is it possible to treat 1 parallel process as no parallel processes or raise error when only one parallel is specified?

Sure, sounds reasonable 👍

> @dreadatour, while you are working on this, could you please also take a look at this example test:
> 
> https://github.com/iterative/datachain/actions/runs/12326251928/job/34406754039?pr=713#step:7:29
> 
> > 862.81s call     tests/examples/test_examples.py::test_get_started_examples[examples/get_started/udfs/parallel.py]
> 
> Does this PR improve that test? Should it take that long?

Nice catch, let me take a look 🙏