@dreadatour 2024-12-14
> @dreadatour, while you are working on this, could you please also take a look at this example test:
> 
> https://github.com/iterative/datachain/actions/runs/12326251928/job/34406754039?pr=713#step:7:29
> 
> > 862.81s call     tests/examples/test_examples.py::test_get_started_examples[examples/get_started/udfs/parallel.py]
> 
> Does this PR improve that test? Should it take that long?

Found an issue. This is because of [this](https://github.com/iterative/datachain/pull/713#discussion_r1884797977) (basically everything runs in single process because batch size is 10k and number of records is 400).

Couple tests:

##### `main` branch

```
$ time python examples/get_started/udfs/parallel.py
Preparing: 400 rows [00:00, 75699.21 rows/s]
Processed: 400 rows [00:35, 11.34 rows/s]
Cleanup: 2 tables [00:00, 10525.23 tables/s]
python examples/get_started/udfs/parallel.py  391.85s user 15.58s system 948% cpu 42.941 total
```

##### this branch

Basically single process

```
$ time python examples/get_started/udfs/parallel.py
Preparing: 400 rows [00:00, 107926.77 rows/s]
Processed: 401 rows [04:01,  1.66 rows/s]
Cleanup: 2 tables [00:00, 10512.04 tables/s]
python examples/get_started/udfs/parallel.py  260.37s user 15.32s system 110% cpu 4:09.27 total
```

##### this branch but batch size is set to `10`

```
$ time python examples/get_started/udfs/parallel.py
Preparing: 400 rows [00:00, 110894.41 rows/s]
Processed: 412 rows [00:38, 10.64 rows/s]
Cleanup: 2 tables [00:00, 10356.31 tables/s]
python examples/get_started/udfs/parallel.py  391.61s user 18.10s system 885% cpu 46.277 total
```