@dreadatour 2024-12-17
> @dreadatour, while you are working on this, could you please also take a look at this example test:
> 
> https://github.com/iterative/datachain/actions/runs/12326251928/job/34406754039?pr=713#step:7:29
> 
> > 862.81s call     tests/examples/test_examples.py::test_get_started_examples[examples/get_started/udfs/parallel.py]
> 
> Does this PR improve that test? Should it take that long?

Fixed: https://github.com/iterative/datachain/actions/runs/12372015291/job/34529325219?pr=713 (409 sec)
Current main for reference: https://github.com/iterative/datachain/actions/runs/12367070520/job/34514761451 (same 410 sec)