@0x2b3bfa0 2024-12-25
Amazing work! 🤩 

Unless I missed something, this sets the foundations for truly scalable distributed processing, where, instead of having a central process batch + pass the rows inefficiently, each worker can query its own batch of records using the same algorithm as `train_test_split` but based on `sys__id`

https://github.com/iterative/datachain/blob/60256d618a79d53c1d18872cd93c6095e442aa1c/src/datachain/toolkit/split.py#L79-L83

Workers would only need to know two variables: the total number of workers, and the current worker's index.