@0x2b3bfa0 2024-12-25
Also, please consider — as a follow-up — making sure that this approach is also efficient on ClickHouse, as mentioned on https://github.com/iterative/datachain/pull/713#discussion_r1890063875